0 deps mode

+++ src/lib/database.ts
@@ -7,7 +7,7 @@ export type Insertable = string | number;
 export interface Database {
     insert(args: Insertable[]): Promise<boolean>;
     update(key: string, to: Insertable, where?: string, whereArgs?: Insertable[]): Promise<boolean>;
-    selectOne<T>(keys: string[], where?: string, whereArgs?: Insertable[]): Promise<T>;
+    selectOne<T>(keys: string[], where?: string, whereArgs?: Insertable[]): Promise<T | null>;
     selectAll<T>(keys: string[], where?: string, whereArgs?: Insertable[]): Promise<T[]>;
     deleteOne(where: string, whereArgs: Insertable[]): Promise<boolean>;
     deleteAll(where: string, whereArgs: Insertable[]): Promise<boolean>;
@@ -17,8 +17,10 @@ export interface Database {
 // some options 
 export class SQLDatabase implements Database {
     tableName: string;
-    database: import('better-sqlite3').Database;
+    // @ts-ignore
+    database: import('better-sqlite3').Database | null;
     keys: string[];
+    // @ts-ignore
     statementCache = new Map<string, import('better-sqlite3').Statement>();
     constructor(options: {
         tableName: string, 
@@ -26,7 +28,11 @@ export class SQLDatabase implements Database {
         keys: string[],
     }) {
         this.tableName = options.tableName;
-        this.database = new (require('better-sqlite3'))(options.file);
+        try {
+            this.database = new (require('better-sqlite3'))(options.file);
+        } catch {
+            this.database = null;
+        }
         this.keys = options.keys;
     }
     prepare(source: string) {
@@ -38,12 +44,14 @@ export class SQLDatabase implements Database {
         return statement;
     }
     insert(args: Insertable[], isReplace = false) {
+        if (!this.database) return Promise.resolve(false);
         const type = isReplace ? 'REPLACE' : 'INSERT';
         const statement = this.prepare(`${type} INTO ${this.tableName} (${this.keys.join(',')}) VALUES (${this.keys.map(() => '?').join(',')})`);
         const result = statement.run(args);
         return Promise.resolve(!!result.changes);
     }
     update(key: string, to: Insertable, where?: string) {
+        if (!this.database) return Promise.resolve(false);
         const statement = this.prepare(
             `UPDATE ${this.tableName} SET ${key} = ?${where ? ` WHERE ${where}` : ""}`
         );
@@ -51,25 +59,30 @@ export class SQLDatabase implements Database {
         return Promise.resolve(!!result.changes);
     }
     selectAll<T>(keys: string[] | string, where?: string, whereArgs?: Insertable[]) {
+        if (!this.database) return Promise.resolve([]);
         const keyList = Array.isArray(keys) ? keys.join(', ') : keys;
         const statement = this.prepare(`SELECT ${keyList} FROM ${this.tableName}${where ? ` WHERE ${where}` : ""}`);
         return Promise.resolve(statement.all(whereArgs || [])) as Promise<T[]>;
     }
     selectOne<T>(keys: string[] | string, where?: string, whereArgs?: Insertable[]) {
+        if (!this.database) return Promise.resolve(null);
         const keyList = Array.isArray(keys) ? keys.join(', ') : keys;
         const statement = this.prepare(`SELECT ${keyList} FROM ${this.tableName}${where ? ` WHERE ${where}` : ""} LIMIT 1`);
-        return Promise.resolve(statement.get(whereArgs || [])) as Promise<T>;
+        return Promise.resolve(statement.get(whereArgs || [])) as Promise<T | null>;
     }
     count(where: string) {
+        if (!this.database) return Promise.resolve(0);
         const statement = this.prepare(`SELECT COUNT(*) FROM ${this.tableName} WHERE ${where}`);
         return Promise.resolve(statement.get()['count(*)'] as number);
     }
     deleteOne(where: string, whereArgs: Insertable[]) {
+        if (!this.database) return Promise.resolve(false);
         const statement = this.prepare(`DELETE FROM ${this.tableName} WHERE ${where} LIMIT 1`);
         const result = statement.run(whereArgs);
         return Promise.resolve(!!result.changes);
     }
     deleteAll(where: string, whereArgs: Insertable[]) {
+        if (!this.database) return Promise.resolve(false);
         const statement = this.prepare(`DELETE FROM ${this.tableName} WHERE ${where} LIMIT 1`);
         const result = statement.run(whereArgs);
         return Promise.resolve(!!result.changes);
